refactor: move TripDetailsPage to component &style
