+++ components/TripDetails/TripDetails.styled.js
