+++ pages/trip/[id].js
@@ -1,22 +1,10 @@
-import {
-  DetailsCard,
-  Header,
-  StyledImage,
-  StyledText,
-  Wrapper,
-} from "../../components/page-styles/TripDetails.styled";
-
-import DeleteButton from "@/components/DeleteButton";
-import EditButton from "@/components/EditButton";
-import Link from "next/link";
-import Location from "@/components/Location";
-import TimePeriod from "@/components/TimePeriod";
 import { useRouter } from "next/router";
 import useSWR from "swr";
+import TripDetails from "@/components/TripDetails";
 
 const fetcher = (...args) => fetch(...args).then((res) => res.json());
 
-export default function TripDetails() {
+export default function TripDetailsPage() {
   const router = useRouter();
   const { id } = router.query;
   const { data: trips, isLoading } = useSWR(
@@ -38,22 +26,13 @@ export default function TripDetails() {
   const { title, location, timePeriod, img } = trips;
 
   return (
-    <main>
-      <Header>
-        <h1>Your Trip</h1>
-        <Link href="/"> ← back </Link>
-      </Header>
-      <StyledImage src={img} width={335} height={300} alt="Dummy Pic" />
-      <DetailsCard>
-        <Location location={location} />
-        <StyledText>{title}</StyledText>
-        <TimePeriod timePeriod={timePeriod} />
-      </DetailsCard>
-      <Wrapper>
-        <EditButton id={id} />
-
-        <DeleteButton onHandleDelete={handleDelete} />
-      </Wrapper>
-    </main>
+    <TripDetails
+      title={title}
+      location={location}
+      handleDelete={handleDelete}
+      img={img}
+      timePeriod={timePeriod}
+      id={id}
+    />
   );
 }
